refactor: split app.ts into multiple modules
- Extract PokerChaseDB to src/db/poker-chase-db.ts
- Extract PokerChaseService to src/services/poker-chase-service.ts
- Extract AggregateEventsStream to src/streams/aggregate-events-stream.ts
- Extract WriteEntityStream to src/streams/write-entity-stream.ts
- Extract ReadEntityStream to src/streams/read-entity-stream.ts
- Keep app.ts as a thin re-export layer for backward compatibility
- Fix TypeScript type errors in test files

This improves code organization and maintainability without changing functionality.
